requirements.txt
yfinance
requests
pandas
numpy
sqlalchemy>=2.0
psycopg[binary]>=3.1
alembic>=1.13
python-dotenv
backoff
